libp2p: Use WindowUpdateMode::OnRead for yamux
Use yamux `WindowUpdateMode::OnRead` for yamux configuration to
enable full Yamux flow control and preventing streams to reset
due to reaching the buffer limit.
The default `WindowUpdateMode::OnReceive` that was in place may have
been overwhelming the receiver if for some reason there was some
slowness attending the buffer.
This avoids issues of peer connections being closed with messages
such as: `buffer of stream grows beyond limit`.
This fixes #566.